Changes to SoundBoard.ps1
Big changes from the original file. Removed a lot of extra code to focus in on getting the media player to work in a job. Got it to work and will play whatever song is tied to the button that is pressed. Next step is to see if I can add volume control the the media player that is running in the job.